refactor: 패키지 분리
domain, util package 추가
