+++ src/main/java/step3/domain/SimulationStep.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 import java.util.Collections;
 import java.util.List;
