+++ src/main/java/step3/view/ResultView.java
@@ -1,7 +1,5 @@
 package step3.view;
 
-import step3.SimulationStep;
-
 import java.io.PrintWriter;
 import java.util.List;
 
@@ -12,10 +10,10 @@ public class ResultView {
         this.output = output;
     }
 
-    public void show(List<SimulationStep> simulationResult) {
+    public void show(List<String> result) {
         output.println("실행 결과");
-        for (int i = 0; i < simulationResult.size(); i++) {
-            output.print(simulationResult.get(i).toString());
+        for (int i = 0; i < result.size(); i++) {
+            output.print(result.get(i));
         }
     }
 }
